ACPs-CA-Server/pyproject.toml
[project]
name = "agent-ca-server"
version = "2.0.0"
description = "Agent CA 认证服务后端 API"
authors = [{ name = "Team", email = "team@ioa.pub" }]
readme = "README.md"
license = "MulanPSL-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.13,<4.0"
dependencies = [
    "alembic>=1.16.1,<2.0.0",
    "cryptography>=45.0.3,<46.0.0",
    "fastapi>=0.115.12,<0.116.0",
    "httpx>=0.28.1,<0.29.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "pydantic>=2.11.5,<3.0.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "sqlalchemy>=2.0.41,<2.1.0",
    "sqlmodel>=0.0.24,<0.1.0",
    "uuid6>=2024.7.10,<2025.0.0",
    "uvicorn[standard]>=0.34.3,<0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0,<26.0.0",
    "coverage>=7.9.1,<8.0.0",
    "flake8>=7.2.0,<8.0.0",
    "pre-commit>=4.2.0,<5.0.0",
    "pytest>=8.4.0,<9.0.0",
    "pytest-asyncio>=1.0.0,<2.0.0",
    "pytest-cov>=6.2.1,<7.0.0",
]

[build-system]
requires = ["poetry-core>=2.1.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Poetry 配置
[tool.poetry]
packages = [{ include = "app" }]
include = [
    { path = "main.py", format = [
        "sdist",
        "wheel",
    ] },
    { path = "alembic.ini", format = [
        "sdist",
        "wheel",
    ] },
    { path = "alembic", format = [
        "sdist",
        "wheel",
    ] },
]

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
pythonpath = [".", "tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "crl: marks tests related to CRL functionality",
    "ocsp: marks tests related to OCSP functionality",
    "crl_ocsp: marks tests for CRL and OCSP integration",
]
